Refactor out editing features functionality
